api/requirements.txt
pytest
flask
Flask-SQLAlchemy
psycopg2-binary
Flask-Login
flask_cors
python-dotenv